+++ src/cli/task-runtime.ts
@@ -1,3 +1,4 @@
+import * as child_process from "child_process";
 import { existsSync, readFileSync, statSync } from "fs";
 import { dirname, join, resolve } from "path";
 import * as path from "path";
@@ -7,7 +8,7 @@ import { $ } from "dax";
 import { PROJEN_DIR, TASKS_MANIFEST_VERSION } from "../common";
 import * as logging from "../logging";
 import type { TasksManifest, TaskSpec, TaskStep } from "../task-model";
-import { systemShell, tool } from "../util/exec";
+import { MAX_BUFFER, tool } from "../util/exec";
 
 // avoids a (false positive) esbuild warning about incorrect imports.
 // `?.default ?? module` keeps this working both as a plain CommonJS require and
@@ -727,4 +728,51 @@ interface ShellOptions {
    * @default "projen"
    */
   readonly shell?: string | string[];
+}
+
+/**
+ * Normalized result of running a command through the system shell (a subset of
+ * node's `SpawnSyncReturns`).
+ */
+export interface SystemShellResult {
+  /** Exit code, or `null` if terminated by a signal or never spawned. */
+  readonly status: number | null;
+  /** Captured stdout, or `null` when stdout was inherited. */
+  readonly stdout: Buffer | null;
+  /** Captured stderr, or `null` when stderr was inherited. */
+  readonly stderr: Buffer | null;
+  /** An error raised while attempting to spawn (not a non-zero exit). */
+  readonly error?: Error;
+}
+
+/**
+ * Runs a command line through the operating system's native shell.
+ *
+ * Calls the OS shell via `child_process` with `shell: true`.
+ * This backs the `system` task shell, which lets a user opt out of the built-in
+ * cross-platform shell and use the host's native shell instead:
+ * `/bin/sh`on POSIX and `cmd.exe` on Windows.
+ *
+ * Never throws for a non-zero exit (that is reported via `status`);
+ * a spawn failure is reported via `error`. The caller supplies the full `env`.
+ */
+export function systemShell(
+  command: string,
+  options: { cwd: string; env?: NodeJS.ProcessEnv; capture?: boolean },
+): SystemShellResult {
+  logging.debug(`${command} (cwd: ${options.cwd})`);
+  const result = child_process.spawnSync(command, {
+    cwd: options.cwd,
+    shell: true,
+    maxBuffer: MAX_BUFFER,
+    env: options.env,
+    // "pipe" for STDERR (when capturing) means it appears in exceptions.
+    stdio: options.capture ? ["inherit", "pipe", "pipe"] : "inherit",
+  });
+  return {
+    status: result.status,
+    stdout: result.stdout ?? null,
+    stderr: result.stderr ?? null,
+    error: result.error,
+  };
 }
